restructure models/ folder
