+++ deeprankcore/Trainer.py
@@ -13,7 +13,7 @@ from torch.nn import MSELoss
 import torch.nn.functional as F
 from torch_geometric.loader import DataLoader
 from deeprankcore.models.metrics import MetricsExporterCollection, MetricsExporter, ConciseOutputExporter
-from deeprankcore.models.community_pooling import community_detection, community_pooling
+from deeprankcore.models.operations.community_pooling import community_detection, community_pooling
 from deeprankcore.domain import targettypes as targets
 
 _log = logging.getLogger(__name__)
