+++ deeprankcore/models/structure/atom.py
@@ -1,0 +1,88 @@
+import numpy
+from enum import Enum
+from deeprankcore.models.structure.residue import Residue
+
+
+class AtomicElement(Enum):
+    "value to represent the type of pdb atoms"
+
+    C = 1
+    O = 2 # noqa: pycodestyle
+    N = 3
+    S = 4
+    P = 5
+    H = 6
+
+    @property
+    def onehot(self) -> numpy.array:
+        value = numpy.zeros(max(el.value for el in AtomicElement))
+        value[self.value - 1] = 1.0
+        return value
+
+
+class Atom:
+    "represents a pdb atom"
+
+    def __init__( # pylint: disable=too-many-arguments
+        self,
+        residue: Residue,
+        name: str,
+        element: AtomicElement,
+        position: numpy.array,
+        occupancy: float,
+    ):
+        """
+        Args:
+            residue(deeprank residue object): the residue that this atom belongs to
+            name(str): pdb atom name
+            element(deeprank atomic element enumeration): the chemical element
+            position(numpy array of length 3): pdb position xyz of this atom
+            occupancy(float): pdb occupancy value
+                            This represents the proportion of structures where the atom is detected at a given position.
+                            Sometimes a single atom can be detected at multiple positions. In that case separate structures exist where sum(occupancy) == 1.
+                            Note that only the highest occupancy atom is used by deeprankcore (see tools.pdb._add_atom_to_residue) 
+        """
+        self._residue = residue
+        self._name = name
+        self._element = element
+        self._position = position
+        self._occupancy = occupancy
+
+    def __eq__(self, other) -> bool:
+        return (
+            isinstance(self, type(other))
+            and self._residue == other._residue
+            and self._name == other._name
+        )
+
+    def __hash__(self) -> hash:
+        return hash((self._residue, self._name))
+
+    def __repr__(self) -> str:
+        return f"{self._residue} {self._name}"
+
+    def change_altloc(self, alternative_atom):
+        "replace the atom's location by another atom's location"
+
+        self._position = alternative_atom.position
+        self._occupancy = alternative_atom.occupancy
+
+    @property
+    def name(self) -> str:
+        return self._name
+
+    @property
+    def element(self) -> AtomicElement:
+        return self._element
+
+    @property
+    def occupancy(self) -> float:
+        return self._occupancy
+
+    @property
+    def position(self) -> numpy.array:
+        return self._position
+
+    @property
+    def residue(self):
+        return self._residue
