+++ deeprankcore/tools/pssm.py
@@ -1,7 +1,8 @@
 from typing import TextIO
-from deeprankcore.models.structure import Residue, Chain
+from deeprankcore.models.structure.pdb_structure import Chain
+from deeprankcore.models.structure.residue import Residue
 from deeprankcore.models.pssm import PssmRow, PssmTable
-from deeprankcore.models.amino_acid import amino_acids
+from deeprankcore.domain.aminoacids import amino_acids
 
 amino_acids_by_letter = {
     amino_acid.one_letter_code: amino_acid for amino_acid in amino_acids
