+++ docs/reference/deeprankcore.models.rst
@@ -52,14 +52,6 @@ deeprankcore.models.metrics module
    :undoc-members:
    :show-inheritance:
 
-deeprankcore.models.pair module
--------------------------------
-
-.. automodule:: deeprankcore.models.pair
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
 deeprankcore.models.polarity module
 -----------------------------------
