+++ tests/domain/test_forcefield.py
@@ -1,8 +1,8 @@
 from pdb2sql import pdb2sql
 
 from deeprankcore.domain.forcefield.parsers import atomic_forcefield
-from deeprankcore.models.pdb import get_structure
-from deeprankcore.models.amino_acid import arginine, glutamate
+from deeprankcore.models.operations.pdb import get_structure
+from deeprankcore.domain.aminoacids import arginine, glutamate
 
 
 def test_atomic_forcefield():
