+++ tests/feature/test_aminoacid.py
@@ -1,10 +1,10 @@
 from pdb2sql import pdb2sql
 
-from deeprankcore.models.amino_acid import serine
-from deeprankcore.models.variant import SingleResidueVariant
+from deeprankcore.domain.aminoacids import serine
+from deeprankcore.models.structure.variant import SingleResidueVariant
 from deeprankcore.features.surfacearea import add_features
 from deeprankcore.models.graph import build_residue_graph
-from deeprankcore.models.pdb import get_structure, get_surrounding_residues
+from deeprankcore.models.operations.pdb import get_structure, get_surrounding_residues
 from deeprankcore.domain import nodefeatures
 from deeprankcore.features.aminoacid import add_features # noqa
