+++ tests/feature/test_atom.py
@@ -3,8 +3,9 @@ from pdb2sql import pdb2sql
 
 from deeprankcore.features.atom import add_features
 from deeprankcore.models.graph import build_atomic_graph
-from deeprankcore.models.structure import Chain, Residue
-from deeprankcore.models.pdb import get_structure, get_surrounding_residues
+from deeprankcore.models.structure.pdb_structure import Chain
+from deeprankcore.models.structure.residue import Residue
+from deeprankcore.models.operations.pdb import get_structure, get_surrounding_residues
 from deeprankcore.domain import nodefeatures
